+++ .storybook/preview.js
@@ -24,6 +24,7 @@ const loaderFn = () => {
         require('../src/stories/Welcome.stories.mdx'),
         require('../src/components/Alert/alert.stories.mdx'),
         require('../src/components/Avatar/avatar.stories.mdx'),
+        require('../src/components/Badge/badge.stories.mdx'),
         require('../src/components/Button/button.stories.mdx'),
         require('../src/components/Input/input.stories.mdx'),
         require('../src/components/AutoComplete/autocomplete.stories.mdx'),