+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## 0.1.6 (2020-05-30)
+## 0.1.7 (2020-05-31)
 
 
 ### Bug Fixes
 
+* **message:** fix message componenet dom remove bug ([547a77c](https://github.com/yehuozhili/bigbear-ui/commit/547a77cc5c0bc9df9fae5ebc354847472ac45f34))
 * ghpage decorator bug ([8e36b73](https://github.com/yehuozhili/bigbear-ui/commit/8e36b7308635b7a4bf9278b7996b9cab8279b1fe))
 
 
@@ -15,6 +16,7 @@
 * **menu:** initial implement menu component ([33a29c7](https://github.com/yehuozhili/bigbear-ui/commit/33a29c77471183a3fcb14ddf742e351576c8d3b0))
 * **menu:** menu component inital complete ([2094eba](https://github.com/yehuozhili/bigbear-ui/commit/2094eba806ead1cb7ea03b8f9aaef9163fe9c0a4))
 * **src:** change style and add alert message component ([21590e7](https://github.com/yehuozhili/bigbear-ui/commit/21590e7ea989c9c2f685a791dcfc26b3cd0239f2))
+* **virtuallist:** complete virtuallist ([8e6b44a](https://github.com/yehuozhili/bigbear-ui/commit/8e6b44a0b9d2ed3d6478d8da6a93f843dc1439c5))